apps/backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-google-vertexai
google-cloud-aiplatform
google-cloud-firestore
pydantic
httpx
python-dotenv
